chore: Restructure Packages
