+++ app/src/main/java/com/hyunjung/todoapp/data/ToDoDao.kt
@@ -5,6 +5,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.hyunjung.todoapp.data.models.ToDoData
 
 @Dao
 interface ToDoDao {
