+++ app/src/main/java/com/hyunjung/todoapp/data/repository/ToDoRepository.kt
@@ -1,6 +1,8 @@
-package com.hyunjung.todoapp.data
+package com.hyunjung.todoapp.data.repository
 
 import androidx.lifecycle.LiveData
+import com.hyunjung.todoapp.data.ToDoDao
+import com.hyunjung.todoapp.data.models.ToDoData
 
 class ToDoRepository(private val toDoDao: ToDoDao) {
